requirements.txt
tornado==4.0.2
torndb==0.3
MySQL-python==1.2.5
PyYAML==3.11
redis==2.10.3
ujson==1.33
requests==2.5.1
